backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "Tiffany Truc", email = "tiffanytruc1@gmail.com" },
  { name = "Jackson Keus", email = "jkeus619@gmail.com" },
  { name = "Jien Zheng", email = "jienho.zheng@gmail.com" },
  { name = "Ella Suchikul", email = "isabella.catrice@gmail.com" },
  { name = "Blaise Duncan", email = "blaiseduncan1@gmail.com" },
  { name = "Mark Canilang", email = "markbcanilang@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "bcrypt>=4.1",
  "bandit>=1.8.3",
  "fastapi[all]",
  "hatch>=1.14.0",
  "passlib[bcrypt]>=1.7.4",
  "pre-commit>=4.1.0",
  "psycopg2-binary>=2.9.10",
  "pydantic-settings>=2.8.1",
  "pyjwt>=2.10.1",
  "python-dotenv>=1.1.0",
  "ruff>=0.9.7",
  "sqlalchemy>=2.0.38",
  "websockets>=15.0.1",
  "pandas>=2.2.3",
]

[project.urls]
Documentation = "https://github.com/SDSU-CompE-561-Spring-2025/leetcode-study-partner#readme"
Issues = "https://github.com/SDSU-CompE-561-Spring-2025/leetcode-study-partner/issues"
Source = "https://github.com/SDSU-CompE-561-Spring-2025/leetcode-study-partner"

[tool.hatch.version]
path = "src/app/__about__.py"

[tool.hatch.envs.default]
# Default environment settings
python = "3.12"
installer = "uv"
path = ".venv"

[tool.hatch.envs.default.scripts]
# Default environment scripts for development and database operations
dev = "fastapi dev {args:src/app/main.py}"
dev_webrtc = "uvicorn src.app.main:app --reload --host 0.0.0.0 --port 8000 --ssl-keyfile ./certs/key.pem --ssl-certfile ./certs/cert.pem"

[tool.hatch.envs.types]
installer = "uv"
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/app tests}"

[tool.coverage.run]
source_pkgs = ["app", "tests"]
branch = true
parallel = true
omit = [
  "src/app/__about__.py",
]

[tool.coverage.paths]
app = ["src/app", "*/app/src/app"]
tests = ["tests", "*/app/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
